API things
You can search location, places and it will autocomplete the search. Display result and you can click 'view details' to be redirected to a Results page. I want to change it so the user needs to press "search" instead of "view details". Also haven't got Results to display on Results page.